Refactored compilation process, so that no copying of files is required after compilation.
More specifically, all ts files are under "src", with the library and tests being in "lib" and "test", respectively.

If this package is installed as a global dependency, the driver will be automatically installed on first use from the command line.

Renamed /models/index.ts to models/TableIndex.ts, so that index.ts can now be used to index all model files. By the same token, renamed the entrypoint to cli.js and library.ts to index.ts.